* 自定义销售报表 由于 float 的精度问题导致导出的 Excel 内包含大于 2 位的小数点，所以改用 Double 类型

+++ core/app/controllers/AnalyzeSkus.java
@@ -89,7 +89,7 @@ public class AnalyzeSkus extends Controller {
     @Check("analyzeskus.skusalesreport")
     public static void processSkuSalesReport(Date from, Date to, String val) {
         try {
-            List<F.T4<String, Long, Long, Float>> sales = OrderItem.querySalesBySkus(from, to, val);
+            List<F.T4<String, Long, Long, Double>> sales = OrderItem.querySalesBySkus(from, to, val);
             render(sales);
         } catch(Exception e) {
             renderJSON(new Ret(Webs.E(e)));
